separate merge operation from code generator file
